Move Instant & Duration into Control.Effect.Time.System.
