+++ fused-effects-time/src/Control/Carrier/Time/System.hs
@@ -7,36 +7,27 @@
 {-# LANGUAGE UndecidableInstances #-}
 module Control.Carrier.Time.System
 ( -- * Time carrier
-  Instant(..)
-, Duration(..)
-, since
+  since
 , time
 , sinceEpoch
 , era
 , runTime
 , TimeC(TimeC)
   -- * Time effect
-, module Control.Effect.Time
+, module Control.Effect.Time.System
 ) where
 
 import Control.Algebra
 import Control.Applicative (Alternative)
 import Control.Carrier.Lift
 import Control.Carrier.Reader
-import Control.Effect.Time
+import Control.Effect.Time.System
 import Control.Monad (MonadPlus)
 import Control.Monad.Fix (MonadFix)
 import Control.Monad.IO.Class (MonadIO)
 import Control.Monad.Trans.Class (MonadTrans(..))
 import Data.Fixed
 import Data.Time.Clock.System
-
-newtype Instant = Instant { getInstant :: SystemTime }
-  deriving (Eq, Ord, Show)
-
-newtype Duration = Duration { getDuration :: Nano }
-  deriving (Eq, Fractional, Num, Ord, Real, Show)
-
 
 since :: Instant -> Instant -> Duration
 since (Instant (MkSystemTime bs bns)) (Instant (MkSystemTime as ans)) = Duration (realToFrac (as - bs) + MkFixed (fromIntegral ans - fromIntegral bns))
